% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base64-image.R
\name{b64_image}
\alias{b64_image}
\title{Return a base64 encoded string of an image (local filesystem or URL)}
\usage{
b64_image(path_or_url)
}
\arguments{
\item{path_or_url}{local path or URL for the image}
}
\description{
Return a base64 encoded string of an image (local filesystem or URL)
}
