% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_attributes.R
\name{bb_attributes}
\alias{bb_attributes}
\title{Parses bitbar attributes into a string.}
\usage{
bb_attributes(...)
}
\arguments{
\item{...}{Attributes to pass. (e.g. href = "https://www.google.com")}
}
\description{
See https://github.com/matryer/bitbar#plugin-api for all possible attributes.
Partial list: href, color, font, size, length, alternate
}
\examples{
bb_attributes(href = "https://www.google.com")
bb_attributes(font = 2 + 2)
bb_attributes(param1 = 14, param2 = "cake")
}
