% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_head.R
\name{bb_head}
\alias{bb_head}
\title{bb_head}
\usage{
bb_head(icon = "\\U0001f4a1", refresh = TRUE)
}
\arguments{
\item{icon}{Description. Defaults to NULL.}

\item{refresh}{Description. Defaults to NULL.}
}
\description{
This is a description.
}
\examples{
bb_head()
}
