% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_inputs.R
\name{bb_inputs}
\alias{bb_inputs}
\title{Make a tibble of options passed to your script through bash.}
\usage{
bb_inputs()
}
\description{
This is a developing feature.
}
\details{
You can pass arguments to your script with \code{'| param1=foo param2=bar'} arguments.
This function collects those inputs when passed to the Rscript you
are running.
}
\examples{
RscriptInputs <- bbInputs()
}
