% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_nest.R
\name{bb_nest}
\alias{bb_nest}
\title{Translate tibble columns into nested submenus.}
\usage{
bb_nest(.data, ...)
}
\arguments{
\item{.data}{A tibble or dataframe.}

\item{...}{Variables to nest into dropdown menus. Each variable forms a new set of dropdowns.}
}
\description{
Translate tibble columns into nested submenus. Each successive variable
will form another dropdown. The final variable is the end-item. Print
with \code{bb_print} or \code{cat}.
}
\examples{
bb_nest(mtcars, cyl, mpg)
mtcars \%>\% rownames_to_column() \%>\% arrange(cyl) \%>\% bb_nest(cyl, rowname)
}
