% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_new.R
\name{bb_new}
\alias{bb_new}
\title{Makes and opens a new script with the BitBar template.}
\usage{
bb_new(name = "myBitBaR", refresh.rate = "1h", dir = getwd())
}
\arguments{
\item{name}{Name of your menu. Isn't used by BitBar so mostly irrelevant. No periods.}

\item{dir}{Where to save the script. Defaults to your working directory.}

\item{refreshRate}{How often would you like the menu to refresh. Format is 10s, 5m, 1h, etc. Defaults to 1h.}
}
\description{
Also runs the command to make the new script executable by BitBar.
}
\examples{
bb_new("MyMenu")
bb_new("MyMenu", "5s", "path/to/my/BitBar/directory")
}
