% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_PopUp.R
\name{bb_PopUp}
\alias{bb_PopUp}
\title{Asks for user input with a generic Apple dialogue box.}
\usage{
bb_PopUp(
  query = "Input below",
  defaultAnswer = "",
  buttonSubmit = "Submit",
  buttonCancel = "Cancel"
)
}
\arguments{
\item{query}{Prompt to display. Defaults to "Input below".}

\item{defaultAnswer}{Initial value in the input box. Defaults to nothing.}

\item{submitButton}{Button to submit input. Defaults to "Submit".}

\item{cancelButton}{Button to cancel. Defaults to "Cancel".}
}
\description{
Note: only works on Apple computers, as BitBar is Apple-only.
}
\examples{
bbPopUp("What is your favorite color?", defaultAnswer = "green")
}
