% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_print.R
\name{bb_print}
\alias{bb_print}
\title{Print strings in BitBar fashion}
\usage{
bb_print(..., nest = 0, sep = "", collapse = "")
}
\arguments{
\item{...}{What to print. Vectors are concatenated.}

\item{nest}{Level of nesting Defaults to 0.}

\item{sep}{What to seperate vectors by. Defaults to "".}

\item{collapse}{What to collapse vectors by. Defaults to "".}
}
\value{
Nothing
}
\description{
I print, BitBar-style.
}
\examples{
bb_print(nest = 1, c("this", "that"))
bb_print(nest = 2, "this", "that")
bb_print(nest = 3, c("this", "that"), "and the other")
bb_print(nest = 4, c("2 + 2 = ", 2 + 2))

}
