% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reachable.R
\name{reachable}
\alias{reachable}
\title{Is the internet accessible?}
\usage{
reachable(fqdn_or_ip = "8.8.8.8")
}
\arguments{
\item{fqdn_or_ip}{FQDN or ip to use for reachability testing. defaults to google dns}
}
\description{
Is the internet accessible?
}
