% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-entry.R
\name{add_entry}
\alias{add_entry}
\alias{add_separator}
\title{Add an entry to the menu}
\usage{
add_entry(text, type = c("direct", "submenu"), level = 1, ...)

add_separator()
}
\arguments{
\item{text}{to display. See \code{type}.}

\item{type}{type of entry.
\itemize{
\item \code{direct} has no dashes prefix and all output goes directly to the menubar. The first use of \code{direct} will set the menubar title.
\item \code{submenu} uses \code{level} repeats of \verb{--} to put these items in submenus
}}

\item{level}{used with \code{submenu} option of \code{type}}

\item{named}{BitBar entry parameters. See \code{References}.}
}
\description{
Use \code{add_separator()} to add a separator line and \code{add_entry()} to make new
menubar entries.
}
\references{
\href{https://github.com/matryer/bitbar#writing-plugins}{BitBar Plugin API Reference}
}
