% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_print.R
\name{bb_print}
\alias{bb_print}
\title{Print text to BitBar.}
\usage{
bb_print(...)
}
\arguments{
\item{...}{What to print. Concatenates arguments without spaces.}
}
\value{
Nothing
}
\description{
Print text to BitBar.
}
