% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitbar-package.R
\docType{package}
\name{bitbar}
\alias{bitbar}
\alias{bitbar-package}
\title{...}
\description{
A good description goes here otherwise CRAN checks fail.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/bitbar}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/bitbar/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
