% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitbar-package.R
\docType{package}
\name{bitbar}
\alias{bitbar}
\alias{bitbar-package}
\title{Craft macOS Menu Bar Applications with Rscript}
\description{
Mat Ryer's macOS utility \url{https://getbitbar.com/} enables any
script to become a menu bar application. Tools are provided to make it
easier to craft these scripts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/bitbar}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/bitbar/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
