% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-bitbar.R
\name{new_bitbar_script}
\alias{new_bitbar_script}
\title{Helper to get started with a new BitBar {bitbar} script}
\usage{
new_bitbar_script(
  save_as,
  title = "Script Title",
  version = "1.0",
  author = Sys.info()[["user"]],
  github_user = author,
  description = "BitBar Plugin Using R",
  dependencies = "R",
  image_url = "",
  about_url = "",
  refresh = ""
)
}
\arguments{
\item{save_as}{full path (including scriptname.R) to where you want the file saved}

\item{title, version, author, github_user, description, dependencies, image_url, about_url}{BitBar metadata passed on to the generator.}

\item{refesh}{if non-blank this will cause the filename to include a refresh hint
in the \code{save_as} filename before the ending \code{.R} so BitBar will run the script
every \code{refresh} period. See \href{https://github.com/matryer/bitbar#configure-the-refresh-time}{Configuring the refresh time}
for more information.}
}
\description{
Helper to get started with a new BitBar {bitbar} script
}
\examples{
if (interactive()) {
  new_bitbar_script(
    save_as = file.path(bitbar_plugins_dir(), "my-awesome-bitbar-script.R"),
    title = ":blue_heart:",
    description = "Example BitBar Script")
  )
}
}
